Moved Converter utilities out of standard namespace package
