+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/itunes/converters/ITunesTopLevelCategoryConverter.kt
@@ -1,10 +1,10 @@
 package com.github.magneticflux.rss.namespaces.itunes.converters
 
+import com.github.magneticflux.rss.children
+import com.github.magneticflux.rss.fallbackPersister
+import com.github.magneticflux.rss.fullName
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesSubCategory
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesTopLevelCategory
-import com.github.magneticflux.rss.namespaces.standard.converters.children
-import com.github.magneticflux.rss.namespaces.standard.converters.fallbackPersister
-import com.github.magneticflux.rss.namespaces.standard.converters.fullName
 import org.simpleframework.xml.convert.Converter
 import org.simpleframework.xml.stream.InputNode
 import org.simpleframework.xml.stream.OutputNode
