+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/standard/converters/RssFeedConverter.kt
@@ -1,5 +1,8 @@
 package com.github.magneticflux.rss.namespaces.standard.converters
 
+import com.github.magneticflux.rss.children
+import com.github.magneticflux.rss.fallbackPersister
+import com.github.magneticflux.rss.fullName
 import com.github.magneticflux.rss.namespaces.standard.elements.Channel
 import com.github.magneticflux.rss.namespaces.standard.elements.Rss
 import org.simpleframework.xml.convert.Converter
