+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/standard/converters/SourceConverter.kt
@@ -1,6 +1,7 @@
 package com.github.magneticflux.rss.namespaces.standard.converters
 
 import com.github.magneticflux.rss.URLTransform
+import com.github.magneticflux.rss.fallbackPersister
 import com.github.magneticflux.rss.namespaces.standard.elements.Source
 import org.simpleframework.xml.convert.Converter
 import org.simpleframework.xml.stream.InputNode
